102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,17 +11,17 @@ int main(void)
 
 	int i;
 
-	unsigned long long fib1 = 1;
+	unsigned long fib1 = 1;
 
-	unsigned long long fib2 = 2;
+	unsigned long fib2 = 2;
 
-	printf("%llu, %llu, ", fib1, fib2);
+	printf("%lu, %lu, ", fib1, fib2);
 
 	for (i = 3; i <= n; i++)
 	{
-		unsigned long long next_fib = fib1 + fib2;
+		unsigned long next_fib = fib1 + fib2;
 
-		printf("%llu, ", next_fib);
+		printf("%lu, ", next_fib);
 		fib1 = fib2;
 		fib2 = next_fib;
 	}
